predict/requirements.txt
albumentations==1.3.0
astroid==2.15.5
black==23.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coverage==7.2.5
cycler==0.11.0
dill==0.3.6
exceptiongroup==1.1.1
filelock==3.9.0
fonttools==4.39.4
idna==3.4
imageio==2.28.1
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
mpmath==1.2.1
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.1
opencv-python-headless==4.7.0.72
packaging==23.1
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.5.1
pluggy==1.0.0
pylint==2.17.4
pyparsing==3.0.9
pytest==7.3.1
pytest-cov==4.0.0
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
typing_extensions==4.5.0
urllib3==1.26.13
wrapt==1.15.0
